Performance improvement for selecting files via BEP53 magnet so parameter
Loop on user selected file indices and then index into torrent.files
instead of looping on torrent.files and checking if the index is in
the index list. Performance-wise, this way is never worse than the
previous implementation and potentially orders of magnitude better.
In the worst case the index list is all the files, so its O(n). In the
best case, the index list is 1, and thus O(1). In constrast, the current
code has a best case of O(n) and a worst case of O(n^2), when the
index list is all the files. For small numbers of files, this is not
noticeable, but can be quite significant for torrents with tens of
thousands of files.